Support disabled Twig autoescape prior to API v3
Based on a template's filesystem path we can determine the origin of a Twig template. Most of the time Twig templates are part of the theme and depending on the theme's API version, we disable autoescape. However, some plugins use Twig templates, too. We can try comparing the template's filesystem path with the paths of all loaded plugins using a ReflectionObject of the plugin's instance (e.g. the 'PicoDeprecated' class is defined in '/plugins/PicoDeprecated/PicoDeprecated.php', thus the plugin's path is '/plugins/PicoDeprecated/'). If all of this fails, we fall back to disabling autoescape, so we don't break BC - breaking BC is up for Pico 3.0.